// lib/enum/app.enum.ts
export enum EToken {
  ACCESS_TOKEN = "accessToken",
  REFRESH_TOKEN = "refreshToken",
  USER_TOKEN = "user",
  USER_ID = "userId",
}

export enum EHttpMethod {
  GET = "GET",
  POST = "POST",
  PUT = "PUT",
  PATCH = "PATCH",
  DELETE = "DELETE",
}

export enum Role {
  USER = "User",
  ADMIN = "Admin",
  ANONYMOUS = "Anonymous",
  MODERATOR = "Moderator",
}

export enum ELandingPage {
  HOME_PAGE = "home-page",
  ABOUT = "about",
  BESPOKE_SHIRTS = "bespoke-shirts",
  BESPOKE_TAILORING = "bespoke-tailoring",
  BESPOKE_TROUSERS = "bespoke-trousers",
  WOMEN_WEAR = "bespoke-womenswear",
  CONTACT_US = "contact-us",
  BOOK_APPOINTMENT = "book-appointment",
  FAQ = "faq",
  MENU_IMAGE = "menu-image",
  RETURN_EXCHANGES = "returns-exchanges",
  LEGAL = "legal",
  FORMAL = "uniform-formal",
  CASUAL = "uniform-casual",
  POLICY = "uniform-policy",
  WEDDING_COLLECTIONS = "wedding-collections",
  WEDDING_GUIDE = "wedding-guide",
  CATEGORY_COLLECTIONS = "category-collections",
}